modify tests for schemes with very big public keys
Our current tests have a failure test case where it generates two key
pairs, signs with one, and tries to veify with the other which should
fail.
This reuires to store two public keys on the stack which won't work for
schemes with very big public keys (like classic OV).

This commit add the option to switch to a low-RAM variant which instead
of generating two key pairs, only generates one and sets the public key
to random bytes before verifying. This variant is behind a flag
"BIG_PUBLIC_KEY_TESTS" which has to be set to the config.mk of the
respective schemes.